requirements.txt
absl-py==2.1.0
agate==1.9.1
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
daff==1.3.46
dbt==1.0.0.38.15
dbt-adapters==1.7.0
dbt-common==1.10.0
dbt-core==1.9.0b2
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.7.3
debugpy==1.8.6
decorator==5.1.1
deepdiff==7.0.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.2
h5py==3.12.1
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.28.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
leather==0.4.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mashumaro==3.13.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
snowplow-tracker==1.0.3
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.3
Telethon==1.37.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
text-unidecode==1.3
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.12
ultralytics-thop==2.0.9
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
